+++ packages/infinity-modules/data-visualizer/src/components/ExportLog.vue
@@ -1,5 +1,5 @@
 <template>
-  <div style="height:100%">
+  <div v-resize="setHeight">
     <v-btn
       small
       color="primary"
@@ -13,7 +13,15 @@
       :headers="headers"
       :items="downloadLogs"
       item-key="_id"
+      fixed-header
+      :height="height"
       :loading="fetchingLogs"
+      :options.sync="options"
+      :itemsPerPage="10"
+      :footer-props="{
+        disableItemsPerPage: true,
+      }"
+      :server-items-length="logCount"
     >
       <!-- eslint-disable-next-line -->
       <template #item.dateFrom="{ item }">
@@ -25,7 +33,7 @@
       </template>
       <!-- eslint-disable-next-line -->
       <template #item.isEmailSent="{ item }">
-        <div>{{ item.status === false ? 'NO' : 'YES' }}</div>
+        <div>{{ item.isEmailSent === false ? 'NO' : 'YES' }}</div>
       </template>
       <!-- eslint-disable-next-line -->
       <template #item.link="{ item }">
@@ -50,6 +58,7 @@ export default {
   data() {
     return {
       timeout: null,
+      options: {},
       headers: [
         { text: 'Requested at', value: 'createdTimestamp' },
         { text: 'Requested by', value: 'createdBy' },
@@ -61,22 +70,29 @@ export default {
         { text: 'Email sent?', value: 'isEmailSent' },
         { text: 'Download link', value: 'link' },
       ],
+      height: window.innerHeight,
     };
   },
-  created() {
-    this.getLog();
-    this.getDownloadLog();
+  mounted() {
+    this.setHeight();
   },
   beforeDestroy() {
     clearTimeout(this.timeout);
   },
   computed: {
-    ...mapState('dataVisualizer', ['downloadLogs', 'fetchingLogs']),
+    ...mapState('dataVisualizer', ['downloadLogs', 'fetchingLogs', 'logCount']),
   },
   methods: {
     ...mapActions('dataVisualizer', ['getDownloadLog']),
+    setHeight() {
+      this.height = window.innerHeight - 248;
+    },
     async getLog() {
-      await this.getDownloadLog();
+      const { page, itemsPerPage } = this.options;
+      await this.getDownloadLog({
+        pageSize: itemsPerPage,
+        pageNumber: page,
+      });
       this.timeout = setTimeout(() => {
         this.getLog();
       }, 120 * 1000);
@@ -92,6 +108,14 @@ export default {
     getDateTo(query) {
       const timestamps = this.parseDateRange(query);
       return formatDate(new Date(timestamps.$lte.$date), 'yyyy-MM-dd HH:mm:ss');
+    },
+  },
+  watch: {
+    options: {
+      deep: true,
+      handler() {
+        this.getLog();
+      },
     },
   },
 };